修改了Redis使用Jackson进行序列化和反序列化时的BUG
