+++ console/resources/goframe/udsp/udsp.config.properties
@@ -120,3 +120,9 @@ clean.realtime.outmoded.cron.expression=0 0 0 * * ?
 ####################### 交互建模的批量作业相关配置 ###############################
 #清空过时批量作业的表达式(默认：0 0 0 * * ?)
 clean.batch.outmoded.cron.expression=0 0 0 * * ?
+
+####################### 清理消费日志相关配置 ###############################
+#清空消费日志的表达式(默认：0 0 0 * * ?)
+clean.consume.log.cron.expression=0 0 0 * * ?
+#保留最近N天的消费日志（默认：100）
+keep.consume.log.period=100
